Turn on -Wextra at "tests", fix some warnings

+++ include/SpTOL.h
@@ -166,8 +166,8 @@ void sptMTTKRP(sptSparseTensor const * const X,
 /**
  * OMP functions
  */
-int sptSparseTensorAddOMP(sptSparseTensor *Y, sptSparseTensor *X, int const nthreads);
-int sptSparseTensorSubOMP(sptSparseTensor *Y, sptSparseTensor *X, int const nthreads);
+int sptSparseTensorAddOMP(sptSparseTensor *Y, sptSparseTensor *X, size_t const nthreads);
+int sptSparseTensorSubOMP(sptSparseTensor *Y, sptSparseTensor *X, size_t const nthreads);
 
 #ifdef __cplusplus
 }
